新绘图框架完全取代旧版绘图框架，使得绘图更加简单、直观、易用。
