+++ framework_common/manshuo_draw/defalut_config.yaml
@@ -51,6 +51,10 @@ basic_set:
   font_title_size: 34
   font_title_color: (0,0,0)
 
+  font_tag: framework_common/manshuo_draw/data/fort/LXGWWenKai-Regular.ttf  #字体路径
+  font_tag_size: 24                     #字体大小
+  font_tag_color: (9, 132, 204)               #字体颜色
+
   is_rounded_corners_img: True      #是否对模块进行圆角处理
   rounded_img_radius: 20            #圆角处理半径
 
@@ -143,8 +147,11 @@ avatar:
   is_shadow_font: False
   software_list:
     - {'type': 'bilibili','background':framework_common/manshuo_draw/data/img/type_software/bili.png,'right_icon':framework_common/manshuo_draw/data/img/type_software/bili_icon.png}
-    - {'type': 'dy','background':framework_common/manshuo_draw/data/img/type_software/dy.png,'right_icon':framework_common/manshuo_draw/data/img/type_software/dy_icon.png}
+    - {'type': 'dy','background':framework_common/manshuo_draw/data/img/type_software/dy.jpeg,'right_icon':framework_common/manshuo_draw/data/img/type_software/dy_icon.png}
     - {'type': 'lu','background':framework_common/manshuo_draw/data/img/type_software/lu.jpg,'right_icon':framework_common/manshuo_draw/data/img/type_software/lu_icon.png}
+    - {'type': 'bangumi','background':framework_common/manshuo_draw/data/img/type_software/bangumi.jpeg,'right_icon':framework_common/manshuo_draw/data/img/type_software/bangumi_icon.png}
+    - {'type': 'xhs','background':framework_common/manshuo_draw/data/img/type_software/xhs.jpeg,'right_icon':framework_common/manshuo_draw/data/img/type_software/xhs_icon.png}
+    - {'type': 'wb','background':,'right_icon':framework_common/manshuo_draw/data/img/type_software/wb_icon.png}
 
 
   common:
@@ -154,7 +161,7 @@ avatar:
     avatar_size: 100
     padding: 30
     padding_up_bottom: 15
-    upshift: 0                     #上偏移量
+    upshift_extra: 0                     #上偏移量
