+++ run/streaming_media/Link_parsing.py
@@ -4,8 +4,9 @@ import json as json_handle
 from developTools.event.events import GroupMessageEvent
 from developTools.message.message_components import Image, File, Video
 from run.streaming_media.service.Link_parsing.core.login_core import ini_login_Link_Prising
-from run.streaming_media.service.Link_parsing.Link_parsing import link_prising, download_video_link_prising
+from run.streaming_media.service.Link_parsing.Link_parsing import download_video_link_prising
 from run.streaming_media.service.Link_parsing.music_link_parsing import netease_music_link_parse
+from run.streaming_media.service.Link_parsing import *
 
 teamlist = {}
